fixed commands add simple-templa:exec command add stub file
